patient-intake-agent/server/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
livekit-api>=0.8.1
python-dotenv>=1.0.1
